Nativize Float#round_with_precision
Gets about 3x speed increases. This isn't a great target for
nativization since it isn't very commonly used, particularly by Rails
itself, but I was desperate to find something besides string inflections
that could usefully be nativized.

N.B. The reason I selected this method for nativization was that the
pure-Ruby implementation does a decent handful of arithmetic operations,
each of which involves instantiating a Numeric and then performing
method calls on it; I figured that doing the math using primitive ints
would be appreciably faster, which apparently it is.